move quality gate to its own package
