+++ src/main/java/edu/hm/hafner/grading/AggregatedScore.java
@@ -11,6 +11,8 @@ import edu.hm.hafner.grading.AnalysisScore.AnalysisScoreBuilder;
 import edu.hm.hafner.grading.CoverageScore.CoverageScoreBuilder;
 import edu.hm.hafner.grading.MetricScore.MetricScoreBuilder;
 import edu.hm.hafner.grading.TestScore.TestScoreBuilder;
+import edu.hm.hafner.qualitygate.QualityGate;
+import edu.hm.hafner.qualitygate.QualityGateResult;
 import edu.hm.hafner.util.FilteredLog;
 import edu.hm.hafner.util.Generated;
